Refactor to reuse the image validation across ipxe and httpboot
