+++ Dockerfile
@@ -14,9 +14,7 @@ RUN go mod download
 # Copy the go source
 COPY cmd/main.go cmd/main.go
 COPY api/ api/
-COPY internal/controller/ internal/controller/
-COPY internal/registry/ internal/registry/
-COPY internal/uki/ internal/uki/
+COPY internal/ internal/
 COPY server/ server/
 COPY templates/ templates/
